Web公開用_Docker版/requirements.txt
Flask
flask-cors
google-cloud-firestore
gunicorn